Refactor packaging to execute on DataflowRunner
