+++ setup.py
@@ -1,0 +1,8 @@
+import setuptools
+
+setuptools.setup(
+    name='data-pipeline',
+    version='0.0.1',
+    packages=setuptools.find_packages(),
+    install_requires=['apache-beam[gcp]']
+)
